# swagger/swagger-output.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Description",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:2222/"
        }
    ],
    "paths": {
...
        "/api/super-id/": {
            "get": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Current users superId",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "User has no superId",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "GET_SUPER_ID_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "SuperId assigned to your account not found"
                                                }
                                            },
                                            "required": ["code", "message"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "code": {
                                                    "type": "string",
                                                    "example": "GET_SUPER_ID_NOT_FOUND"
                                                },
                                                "message": {
                                                    "type": "string",
                                                    "example": "SuperId assigned to your account not found"
                                                }
                                            },
                                            "required": ["code", "message"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Created superId",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "validTo": {
                                        "type": "number"
                                    },
                                    "pin": {
                                        "type": "string"
                                    },
                                    "studentId": {
                                        "type": "string"
                                    },
                                    "identityCardId": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Updated superId",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "superId": {
                                                    "$ref": "#/components/schemas/SuperId"
                                                }
                                            },
                                            "required": ["superId"]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "identityCardId": {
                                        "example": "any"
                                    },
                                    "pin": {
                                        "example": "any"
                                    },
                                    "studentId": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/super-id/challenge/": {
            "post": {
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Challenge response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "authorizations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AuthorizationExtendedFacilities"
                                                    }
                                                },
                                                "user": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "hello@kzakrzewski.pl"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "VISITOR"
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "email",
                                                        "role"
                                                    ]
                                                },
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": [
                                                "authorizations",
                                                "user"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "serverTime": {
                                            "type": "string",
                                            "example": "1"
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "authorizations": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AuthorizationExtendedFacilities"
                                                    }
                                                },
                                                "user": {
                                                    "type": "object",
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "example": "1"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "hello@kzakrzewski.pl"
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "VISITOR"
                                                        }
                                                    },
                                                    "required": [
                                                        "id",
                                                        "email",
                                                        "role"
                                                    ]
                                                },
                                                "profile": {
                                                    "$ref": "#/components/schemas/Profile"
                                                }
                                            },
                                            "required": [
                                                "authorizations",
                                                "user"
                                            ]
                                        }
                                    },
                                    "required": [
                                        "status",
                                        "serverTime",
                                        "data"
                                    ],
                                    "xml": {
                                        "name": "main"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "challengeId": {
                                        "example": "any"
                                    },
                                    "pin": {
                                        "example": "any"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facility": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Facility name"
                    },
                    "occupancy": {
                        "type": "array",
                        "items": {
                            "xml": {
                                "name": "facilityoccupancy"
                            },
                            "$ref": "#/components/schemas/FacilityOccupancy"
                        }
                    }
                },
                "required": ["id", "name", "occupancy"],
                "xml": {
                    "name": "Facility"
                }
            },
...
            "Profile": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "firstName": {
                        "type": "string",
                        "example": "Kacper"
                    },
                    "surname": {
                        "type": "string",
                        "example": "Zakrzewski"
                    },
                    "description": {
                        "type": "string",
                        "example": "I'm a programmer"
                    },
                    "userId": {
                        "type": "string",
                        "example": "1"
                    },
                    "avatarUrl": {
                        "type": "string",
                        "example": "https://kzakrzewski.pl/avatar.png"
                    }
                },
                "required": [
                    "id",
                    "firstName",
                    "surname",
                    "description",
                    "userId"
                ],
                "xml": {
                    "name": "Profile"
                }
            },
...
            "AuthorizationExtendedFacilities": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "scopeFacility": {
                        "type": "array",
                        "items": {
                            "xml": {
                                "name": "facility"
                            },
                            "$ref": "#/components/schemas/Facility"
                        }
                    },
                    "assignedUsers": {
                        "type": "array",
                        "example": ["1", "2", "3"],
                        "items": {
                            "type": "string"
                        }
                    },
                    "createdAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "expireAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "createdByUserId": {
                        "type": "string",
                        "example": "1"
                    }
                },
                "required": [
                    "id",
                    "scopeFacility",
                    "assignedUsers",
                    "createdAt",
                    "expireAt",
                    "createdByUserId"
                ],
                "xml": {
                    "name": "AuthorizationExtendedFacilities"
                }
            },
            "SuperId": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "pin": {
                        "type": "string",
                        "example": "1234"
                    },
                    "secret": {
                        "type": "string",
                        "example": "=SID==my-little-secret"
                    },
                    "studentId": {
                        "type": "string",
                        "example": "1"
                    },
                    "identityCardId": {
                        "type": "string",
                        "example": "1"
                    },
                    "userId": {
                        "type": "string",
                        "example": "1"
                    },
                    "validTo": {
                        "type": "number",
                        "example": 1577836800000
                    }
                },
                "required": ["id", "secret", "userId", "validTo"],
                "xml": {
                    "name": "SuperId"
                }
            },
...
            "FacilityOccupancy": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1"
                    },
                    "isOccupied": {
                        "type": "boolean",
                        "example": true
                    },
                    "facilityId": {
                        "type": "string",
                        "example": "1"
                    },
                    "facility": {
                        "xml": {
                            "name": "facility"
                        },
                        "$ref": "#/components/schemas/Facility"
                    },
                    "createdAt": {
                        "type": "number",
                        "example": 1577836800000
                    },
                    "relatedEntryId": {
                        "type": "string",
                        "example": "2"
                    }
                },
                "required": [
                    "id",
                    "isOccupied",
                    "facilityId",
                    "facility",
                    "createdAt",
                    "relatedEntryId"
                ],
                "xml": {
                    "name": "FacilityOccupancy"
                }
            },
...
        }
    }
}